% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplots_tikka11124.R
\name{boxplots}
\alias{boxplots}
\title{Boxplottings}
\usage{
boxplots(tvt, Group, Outcome, Out, oute, other)
}
\arguments{
\item{tvt}{The dataset}
}
\value{
the plot
}
\description{
Converts data to boxplots.
Written by Pauli Tikka. University of Turku. Updated 1.11.24.
}
\examples{
temp1 <- boxplots;
}
